backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
anthropic>=0.40.0
supabase>=2.10.0
openai>=1.50.0
google-genai>=1.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
httpx>=0.27.0
